Add scopes and symbol classes, modify listeners
The base listener implementation DefPhaseSimpleJava adds a few types-checking
on variables declaration, annotates the parse tree with the type then saves on
the scope.

The BaseScope class has methods to add symbols to scopes and resolve them, both
locally and in the enclosing scope.

As for the symbol package, it contains implementation for variable and function
symbols.