Fix a bug in the DeclFuncoes methods
Performs a proper check in the function declaration and reports errors
on collision on the name.

Creates a new class for Compile Errors, such as, overloading errors on
function names.

Implements a remove function, which, delete scopes with errors on
overloading.

Adds a few extra tests to cover these modifications.